requirements.txt
# Web Framework
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.30
alembic>=1.13.2
psycopg[binary]>=3.2.1

# Data Validation / Settings
pydantic>=2.7
pydantic-settings>=2.3
email-validator>=2.1.1

# Authentication & Security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.3
slowapi>=0.1.9

# Encryption
cryptography>=44.0

# Env
python-dotenv>=1.0.1

# Date/Time
python-dateutil>=2.9.0.post0

# HTTP Client
httpx>=0.27

# WebSocket Support
websockets>=12.0

# ML Model Runtime
scikit-learn==1.8.0

# Document Extraction (Gemini LLM + PDF parsing)
google-generativeai>=0.8.0
PyPDF2>=3.0

# Testing
pytest>=8.2
pytest-asyncio>=0.23.8
pytest-cov>=5.0

# Dev Tools (optional)
black>=24.4
isort>=5.13
flake8>=7.1

# Logging
structlog>=24.2
